# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Aquachile",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/monthly-pp": {
            "get": {
                "tags": [
                    "KPI Monthly Programming Progress"
                ],
                "summary": "Show all monthly programming progress kpi",
                "description": "Returns a list of monthly programming progress kpi",
                "operationId": "3439d1c332428f98185bc1cdaca71658",
                "responses": {
                    "200": {
                        "description": "Show all monthly programming progress kpi."
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/monthly-pp-line/{id}": {
            "get": {
                "tags": [
                    "KPI Monthly Programming Progress"
                ],
                "summary": "Show all monthly programming progress kpi by line",
                "description": "Returns a list of monthly programming progress kpi by line",
                "operationId": "0150d81604fb730adc6e0dcc52c97bae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Line ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show all monthly programming progress kpi by line."
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/monthly-pp-shift/{id}": {
            "get": {
                "tags": [
                    "KPI Monthly Programming Progress"
                ],
                "summary": "Show all monthly programming progress kpi by shift",
                "description": "Returns a list of monthly programming progress kpi by shift",
                "operationId": "7e78a041d4fc152b61686e957b9dbd67",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Shift ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show all monthly programming progress kpi by shift."
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/monthly-pp/{id}": {
            "post": {
                "tags": [
                    "KPI Monthly Programming Progress"
                ],
                "summary": "Create a new monthly programming progress KPI",
                "description": "Creates a new monthly programming progress KPI with the received data",
                "operationId": "de4d6caee1bebb612f69dda683dee8e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Line ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "monthly programming progress KPI object that needs to be created",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "monthly_order",
                                    "line_name"
                                ],
                                "properties": {
                                    "monthly_order": {
                                        "description": "The monthly programming order",
                                        "type": "integer"
                                    },
                                    "uuid": {
                                        "description": "The shift uuid",
                                        "type": "uuid"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "monthly_order": 40000,
                                    "uuid": "123e4567-e89b-12d3-a456-426614174000"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "default": {
                        "description": "An error occurred."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}